Major changes to lit.
Lit no longer comes bundled with a version of lua, you must now install
it yourself. However, the lua file that gets created (bin/lit) is now a
standalone executable which can be in any directory. This opens up many
more options for using lit. It can be easily installed by a package
manager, and it can be made into a binary using srlua.