Check for match before accessing it
So at least for the current dlang regex library, if you access `matches[...]` but no match was actually found, you'll get random data (from previous matches) that later causes crashes when indexing `codeLinenums`. To circumvent this we can simply check for `matches`, which should be backwards-compatible with older regex versions.